+++ CHANGELOG.md
@@ -1,13 +1,12 @@
 ### Version 0.6.1
 - [chore] update packages & cleanup & work on improving documentation
+- [express-node] add graphql sample code for simple CRUD (with graphiql queries & variables provided)
 - IN PROGRESS
   - [chore] disable console.log in production / proper console logging
     - https://stackoverflow.com/questions/7042611/override-console-log-for-production/21789576
     - https://ourcodeworld.com/articles/read/104/how-to-override-the-console-methods-in-javascript
     - https://gist.github.com/pincheira/2724082
     - https://stackoverflow.com/questions/63075459/override-console-log-method-without-losing-the-original-stack-javascript
-  - [express-node] add graphql sample code for simple CRUD
-  - [chore] folder reorganisation, fix documents, re-test and re-document deployments, clean up GCP cloud documentation... 
 
 ### Version 0.6.0
 - [chore] update packages & cleanup & work on improving documentation